+++ components/hero-section.tsx
@@ -7,13 +7,7 @@ import { useTikTokAnalysis } from "@/lib/hooks/use-tiktok-analysis";
 import { useSaveTikTokAnalysisWithCredibility } from "@/lib/hooks/use-saved-analyses";
 import { useAnimatedProgress } from "@/lib/hooks/use-animated-progress";
 import { toast } from "sonner";
-import { AnalysisRenderer } from "@/components/analysis-renderer";
 import { useGlobalTranslation } from "@/components/global-translation-provider";
-import { OriginTracingDiagram } from "@/components/analysis/origin-tracing-diagram";
-import { SentimentDisplay } from "@/components/analysis/sentiment-display";
-import { useDiagramExpansion } from "@/lib/hooks/useDiagramExpansion";
-import { PoliticalBiasMeter } from "@/components/ui/political-bias-meter";
-import Link from "next/link";
 import { useAuth } from "@/lib/hooks/use-auth";
 import {
   UrlInputForm,
